--- cam_utils.py
@@ -1,14 +1,0 @@
-import cv2
-import numpy as np
-import torch
-
-def visualize_cam(mask, img):
-    heatmap = cv2.applyColorMap(np.uint8(255 * mask.squeeze()), cv2.COLORMAP_JET)
-    heatmap = torch.from_numpy(heatmap).permute(2, 0, 1).float().div(255)
-    b, g, r = heatmap.split(1)
-    heatmap = torch.cat([r, g, b])
-    
-    result = heatmap+img.cpu()
-    result = result.div(result.max()).squeeze()
-    
-    return heatmap, result
