--- gradcam.py
@@ -1,64 +1,0 @@
-import torch
-import torch.nn.functional as F
-import numpy as np
-
-class Grad_CAM(object):
-    def __init__(self, model_dict, verbose=False):
-        model_type = model_dict['type']
-        layer_name = model_dict['layer_name']
-        self.model_arch = model_dict['arch']
-
-        self.gradients = dict()
-        self.activations = dict()
-        def backward_hook(module, grad_input, grad_output):
-            self.gradients['value'] = grad_output[0]
-            return None
-        def forward_hook(module, input, output):
-            self.activations['value'] = output
-            return None
-        target_layer = layer_name
-        target_layer.register_forward_hook(forward_hook)
-        target_layer.register_backward_hook(backward_hook)
-
-        if verbose:
-            try:
-                input_size = model_dict['input_size']
-            except KeyError:
-                print("please specify size of input image in model_dict. e.g. {'input_size':(1024, 1024)}")
-                pass
-            else:
-                device = 'cuda' if next(self.model_arch.parameters()).is_cuda else 'cpu'
-                self.model_arch(torch.zeros(1, 3, *(input_size), device=device))
-                print('saliency_map size :', self.activations['value'].shape[2:])
-
-
-    def forward(self, input, class_idx=None, retain_graph=False):
-        b, c, h, w = input.size()
-
-        output = self.model_arch(input)
-        logit = output[1]
-        if class_idx is None:
-            score = logit[:, logit.max(1)[-1]].squeeze()
-        else:
-            score = logit[:, class_idx].squeeze()
-
-        self.model_arch.zero_grad()
-        score.backward(retain_graph=retain_graph)
-        gradients = self.gradients['value']
-        activations = self.activations['value']
-        b, k, u, v = gradients.size()
-
-        alpha = gradients.view(b, k, -1).mean(2)
-        #alpha = F.relu(gradients.view(b, k, -1)).mean(2)
-        weights = alpha.view(b, k, 1, 1)
-
-        saliency_map = (weights*activations).sum(1, keepdim=True)
-        saliency_map = F.relu(saliency_map)
-        saliency_map = F.upsample(saliency_map, size=(h, w), mode='bilinear', align_corners=False)
-        saliency_map_min, saliency_map_max = saliency_map.min(), saliency_map.max()
-        saliency_map = (saliency_map - saliency_map_min).div(saliency_map_max - saliency_map_min).data
-
-        return saliency_map, logit
-
-    def __call__(self, input, class_idx=None, retain_graph=False):
-        return self.forward(input, class_idx, retain_graph)
